Deck
A class creating a 'Deck' of the 'Cards' by placing them within an array then divid them according to their faces, suits and numbers. Within also contain methods in which can create a new deck and shuffle as well as the showing and drawing of the cards from and their reintroduction to the deck.